Rename: page layout 컴포넌트 pages 하위로 이동
