+++ src/routes/index.tsx
@@ -1,9 +1,9 @@
 import { RouterProvider, createBrowserRouter } from 'react-router-dom'
 
-import { MainLayout } from '@/components/features/Layout/MainLayout'
-import { ProfileQuestionLayout } from '@/components/features/Layout/ProfileQuestionLayout'
 import ErrorPage from '@/pages/ErrorPage'
 import GroupPage from '@/pages/GroupPage'
+import { MainLayout } from '@/pages/Layout/MainLayout'
+import { ProfileQuestionLayout } from '@/pages/Layout/ProfileQuestionLayout'
 import LoginPage from '@/pages/LoginPage'
 import LoginRedirectPage from '@/pages/LoginRedirectPage'
 import MainPage from '@/pages/MainPage'
